requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
psycopg2-binary>=2.9
djangorestframework~=3.14.0